test: Add tests in inital test file test_linkedlist.py #10
